Small changes and adaptions

+++ src/main/java/ru/mousecray/endmagic/init/EMEvents.java
@@ -59,7 +59,8 @@ import java.util.Arrays;
 import java.util.Optional;
 import java.util.Random;
 
-import static ru.mousecray.endmagic.init.EMBlocks.*;
+import static ru.mousecray.endmagic.init.EMBlocks.enderLeaves;
+import static ru.mousecray.endmagic.init.EMBlocks.enderLog;
 import static ru.mousecray.endmagic.network.PacketTypes.UPDATE_COMPAS_TARGET;
 import static ru.mousecray.endmagic.network.PacketTypes.UPDATE_PHANROM_AVOIDINCAPABILITY;
 import static ru.mousecray.endmagic.tileentity.TilePhantomAvoidingBlockBase.maxAvoidTicks;
@@ -68,6 +69,14 @@ import static ru.mousecray.endmagic.worldgen.trees.WorldGenPhantomTree.areaRequi
 
 @EventBusSubscriber(modid = EM.ID)
 public class EMEvents {
+
+    private static final int teleportRadius = 40;
+    private static ImmutableMap<Block, Item> coal2diamond = ImmutableMap.of(
+            EMBlocks.dragonCoal, EMItems.dragonDiamond,
+            EMBlocks.immortalCoal, EMItems.naturalDiamond,
+            EMBlocks.naturalCoal, EMItems.phantomDiamond,
+            EMBlocks.phantomCoal, EMItems.immortalDiamond
+    );
 
     @SubscribeEvent
     public static void onCapaAttachToWorld(AttachCapabilitiesEvent<World> event) {
@@ -79,7 +88,8 @@ public class EMEvents {
         World world = event.getEntityPlayer().world;
         BlockPos pos = event.getPos();
         IBlockState blockState = world.getBlockState(pos);
-        if (/*!event.getEntityPlayer().world.isRemote && */(blockState.getBlock() == enderLog || blockState.getBlock() == enderLeaves) && blockState.getValue(enderLog.blockType) == EnderBlockTypes.EnderTreeType.PHANTOM) {
+        if (/*!event.getEntityPlayer().world.isRemote && */(blockState.getBlock() == enderLog || blockState.getBlock() == enderLeaves) &&
+                blockState.getValue(enderLog.blockType) == EnderBlockTypes.EnderTreeType.PHANTOM) {
             PhantomAvoidingGroupCapability capability = world.getCapability(PhantomAvoidingGroupCapabilityProvider.avoidingGroupCapability, null);
             if (capability != null) {
                 PhantomAvoidingGroup tree = capability.groupAtPos.get(event.getPos());
@@ -104,9 +114,7 @@ public class EMEvents {
             BlockPos saplingPos = pos.subtract(((TilePhantomAvoidingBlockBase) tileEntity).offsetFromSapling);
             WorldGenUtils.generateInArea(saplingPos.add(areaRequirementsMin), saplingPos.add(areaRequirementsMax), p -> {
                 TileEntity tileEntity1 = world.getTileEntity(p);
-                if (tileEntity1 instanceof TilePhantomAvoidingBlockBase) {
-                    tree.blocks.add(p.toImmutable());
-                }
+                if (tileEntity1 instanceof TilePhantomAvoidingBlockBase) tree.blocks.add(p.toImmutable());
             });
         }
     }
@@ -123,24 +131,21 @@ public class EMEvents {
             capability.forRemove.clear();
             capability.forAdded.clear();
 
-            for (PhantomAvoidingGroup group : capability.allGroups) {
+            for (PhantomAvoidingGroup group : capability.allGroups)
                 if (group.avoidingStarted) {
 
-                    if (group.increment > 0) {
-                        if (group.avoidTicks >= maxAvoidTicks) {
-                            if (!world.isRemote)
-                                teleportTree(world, group, capability);
-                            capability.forRemove.add(group);
-                            group.avoidingStarted = false;
-                        } else
-                            group.avoidTicks += group.increment;
-                    } else if (group.increment < 0) {
-                        if (group.avoidTicks <= 0) {
-                            capability.forRemove.add(group);
-                            group.avoidingStarted = false;
-                        } else
-                            group.avoidTicks += group.increment;
-                    }
+                    if (group.increment > 0) if (group.avoidTicks >= maxAvoidTicks) {
+                        if (!world.isRemote)
+                            teleportTree(world, group, capability);
+                        capability.forRemove.add(group);
+                        group.avoidingStarted = false;
+                    } else
+                        group.avoidTicks += group.increment;
+                    else if (group.increment < 0) if (group.avoidTicks <= 0) {
+                        capability.forRemove.add(group);
+                        group.avoidingStarted = false;
+                    } else
+                        group.avoidTicks += group.increment;
                     for (BlockPos pos : group.blocks) {
                         TileEntity tileEntity = world.getTileEntity(pos);
                         if (tileEntity instanceof TilePhantomAvoidingBlockBase) {
@@ -150,20 +155,17 @@ public class EMEvents {
                         }
                     }
                 }
-            }
         }
     }
 
-    private static final int teleportRadius = 40;
-
     private static void teleportTree(World world, PhantomAvoidingGroup group, PhantomAvoidingGroupCapability capability) {
-        Optional<TilePhantomAvoidingBlockBase> anyTile = group.blocks.stream().map(world::getTileEntity).filter(t -> t instanceof TilePhantomAvoidingBlockBase).map(t -> (TilePhantomAvoidingBlockBase) t).findAny();
+        Optional<TilePhantomAvoidingBlockBase> anyTile = group.blocks.stream().map(world::getTileEntity).filter(t -> t instanceof TilePhantomAvoidingBlockBase).map(
+                t -> (TilePhantomAvoidingBlockBase) t).findAny();
         anyTile.ifPresent(t -> {
             BlockPos saplingPos = t.getPos().subtract(t.offsetFromSapling);
             BlockPos newSaplingPos;
-            do {
-                newSaplingPos = world.getTopSolidOrLiquidBlock(saplingPos.add(world.rand.nextInt(2 * teleportRadius) - teleportRadius, 0, world.rand.nextInt(2 * teleportRadius) - teleportRadius));
-            } while (newSaplingPos.getY() == -1);
+            do newSaplingPos = world.getTopSolidOrLiquidBlock(saplingPos.add(world.rand.nextInt(2 * teleportRadius) - teleportRadius, 0, world.rand.nextInt(2 * teleportRadius) - teleportRadius));
+            while (newSaplingPos.getY() == -1);
             System.out.println(world.getBlockState(newSaplingPos).getBlock());
 
             BlockPos teleportOffset = newSaplingPos.subtract(saplingPos);
@@ -210,24 +212,21 @@ public class EMEvents {
     @SubscribeEvent
     public static void onPlayerEnter(EntityJoinWorldEvent event) {
         if (!event.getWorld().isRemote)
-            if (event.getEntity() instanceof EntityPlayer) {
-                Optional.ofNullable(((WorldServer) event.getWorld()).getChunkProvider()
-                        .getNearestStructurePos(event.getWorld(), "Stronghold", new BlockPos(event.getEntity()), false))
-                        .map(pos ->
-                                UPDATE_COMPAS_TARGET.packet()
-                                        .writeInt(0)
-                                        .writePos(pos))
-                        .ifPresent(p -> p.sendToPlayer((EntityPlayer) event.getEntity()));
-            }
+            if (event.getEntity() instanceof EntityPlayer) Optional.ofNullable(((WorldServer) event.getWorld()).getChunkProvider()
+                    .getNearestStructurePos(event.getWorld(), "Stronghold", new BlockPos(event.getEntity()), false))
+                    .map(pos ->
+                            UPDATE_COMPAS_TARGET.packet()
+                                    .writeInt(0)
+                                    .writePos(pos))
+                    .ifPresent(p -> p.sendToPlayer((EntityPlayer) event.getEntity()));
     }
 
     @SideOnly(Side.CLIENT)
     //    @SubscribeEvent
     public static void loadLastWorld(GuiOpenEvent event) {
         Minecraft mc = Minecraft.getMinecraft();
-        if (event.getGui() instanceof GuiMainMenu) {
-            mc.displayGuiScreen(new GuiWorldSelection((GuiMainMenu) event.getGui()));
-        } else if (event.getGui() instanceof GuiWorldSelection) {
+        if (event.getGui() instanceof GuiMainMenu) mc.displayGuiScreen(new GuiWorldSelection((GuiMainMenu) event.getGui()));
+        else if (event.getGui() instanceof GuiWorldSelection) {
             GuiListWorldSelection guiListWorldSelection = new GuiListWorldSelection((GuiWorldSelection) event.getGui(), mc, 100, 100, 32, 100 - 64, 36);
             try {
                 guiListWorldSelection.getListEntry(0).joinWorld();
@@ -238,55 +237,36 @@ public class EMEvents {
 
     //TODO: onUseBonemeal
     @SubscribeEvent
-    public static void onUseBonemeal(BonemealEvent event)
-    {
-        World world  = event.getWorld();
+    public static void onUseBonemeal(BonemealEvent event) {
+        World world = event.getWorld();
         BlockPos pos = event.getPos();
         Random rand = event.getEntityPlayer().getRNG();
-        if(event.getBlock().getBlock() instanceof IEndSoil) //If it is Soil
+        if (event.getBlock().getBlock() instanceof IEndSoil) //If it is Soil
         {
-            IEndSoil soil = (IEndSoil)event.getBlock().getBlock();
-            if(soil.canUseBonemeal()) //if we can use bone meal
+            IEndSoil soil = (IEndSoil) event.getBlock().getBlock();
+            if (soil.canUseBonemeal()) //if we can use bone meal
             {
                 for (int x = -2; x < 2; ++x) //process neighboring blocks
-                {
                     for (int z = -2; z < 2; ++z)//process neighboring blocks
                     {
                         BlockPos pos2 = pos.add(x, 0, z);
-                        if(EMUtils.isSoil(world.getBlockState(pos2), true, true, EndSoilType.DIRT, EndSoilType.GRASS))
-                        {
+                        //TODO: add custom end grass and remove STONE from this
+                        EMUtils.getBonemealCropsFromType(world.getBlockState(pos2), rand, event.getEntityPlayer(), EndSoilType.STONE, EndSoilType.DIRT, EndSoilType.GRASS)
+                        if () {
                             int chance = rand.nextInt(1000) + 1; //create random
-                            Block block2 =  world.getBlockState(pos2).getBlock();
-                            if (block2 instanceof IEndSoil) {
-                                if (world.isAirBlock(pos2.up())) {
-                                    if (block2 instanceof IEndSoil)
+                            IBlockState plant = EMUtils.getBonemealCrops(world.getBlockState())
+                            Block block2 = world.getBlockState(pos2).getBlock();
+                            if (block2 instanceof IEndSoil) if (world.isAirBlock(pos2.up())) {
+                                if (block2 instanceof IEndSoil)
                                     world.setBlockState(pos2.up(), soil.getBonemealCrops(rand, event.getEntityPlayer(), block2.getDefaultState()));
-                                    ItemDye.spawnBonemealParticles(world, pos2, 5);
-                                }
+                                ItemDye.spawnBonemealParticles(world, pos2, 5);
                             }
                         }
                     }
-                }
                 event.setResult(Event.Result.ALLOW);
             }
         }
-        else if (event.getBlock().getBlock() == Blocks.END_STONE)
-        {
-            for (int x = -2; x < 2; ++x)
-            {
-                for (int z = -2; z < 2; ++z)
-                {
-                    if (world.isAirBlock(pos.add(x, 1, z)) && world.getBlockState(pos.add(x, 0, z)).getBlock() == Blocks.END_STONE && event.getEntityPlayer().getRNG().nextInt(10) > 8)
-                    {
-                        world.setBlockState(pos.add(x, 1, z), EMBlocks.enderTallgrass.getDefaultState());
-                        ItemDye.spawnBonemealParticles(world, pos, 5);
-                    }
-                }
-            }
-            event.setResult(Event.Result.ALLOW);
-        }
     }
-
 
     @SubscribeEvent
     public static void onPressureExplosionCoal(ExplosionEvent.Start event) {
@@ -300,13 +280,6 @@ public class EMEvents {
                 })
                 .ifPresent(world::spawnEntity);
     }
-
-    private static ImmutableMap<Block, Item> coal2diamond = ImmutableMap.of(
-            EMBlocks.dragonCoal, EMItems.dragonDiamond,
-            EMBlocks.immortalCoal, EMItems.naturalDiamond,
-            EMBlocks.naturalCoal, EMItems.phantomDiamond,
-            EMBlocks.phantomCoal, EMItems.immortalDiamond
-    );
 
     private static Optional<EntityItem> getDiamond(World world, BlockPos i) {
         return Optional.ofNullable(coal2diamond.get(world.getBlockState(i).getBlock()))
@@ -388,14 +361,13 @@ public class EMEvents {
                     world.spawnEntity(entityarrow);
                 }
 
-                world.playSound((EntityPlayer) null, player.posX, player.posY, player.posZ, SoundEvents.ENTITY_ARROW_SHOOT, SoundCategory.PLAYERS, 1.0F, 1.0F / (new Random().nextFloat() * 0.4F + 1.2F) + f * 0.5F);
+                world.playSound((EntityPlayer) null, player.posX, player.posY, player.posZ, SoundEvents.ENTITY_ARROW_SHOOT, SoundCategory.PLAYERS, 1.0F,
+                        1.0F / (new Random().nextFloat() * 0.4F + 1.2F) + f * 0.5F);
 
                 if (!flag1 && !player.capabilities.isCreativeMode) {
                     stack.shrink(1);
 
-                    if (stack.isEmpty()) {
-                        player.inventory.deleteStack(stack);
-                    }
+                    if (stack.isEmpty()) player.inventory.deleteStack(stack);
                 }
 
                 player.addStat(StatList.getObjectUseStats(stack.getItem()));
